Add functionality to import nPYc object directly
