+++ pyIstocsy/_utilities.py
@@ -9,8 +9,6 @@ Created on Fri Apr  6 11:32:54 2018
 """
 import numpy as np
 import pandas
-from matplotlib import pyplot as plt
-from matplotlib.colors import LinearSegmentedColormap
 from scipy.stats import pearsonr, spearmanr, kendalltau
 from statsmodels.stats.multitest import multipletests
 import networkx as nx
@@ -20,15 +18,6 @@ def _loadCSV(self, intensityDataFile, featureMetadataFile):
 
 	if ((intensityDataFile is None) or (featureMetadataFile is None)):
 		raise TypeError('intensityDataFile and featureMetadataFile must be set')
-
-	# Create dataset
-	class Dataset(object):
-
-		def __init__(self):
-			self.intensityData = np.array(None)
-			self.featureMetadata = pandas.DataFrame(None, columns=['Feature Name', 'Retention Time', 'm/z'])
-
-	self.dataset = Dataset()
 
 	self.dataset.intensityData = np.genfromtxt(intensityDataFile, delimiter=',')
 	self.dataset.featureMetadata = pandas.read_csv(featureMetadataFile)
@@ -40,37 +29,18 @@ def _loadCSV(self, intensityDataFile, featureMetadataFile):
 		raise ValueError('intensityData and featureMetadata have different dimensions')
 
 
-#def _loadData(self):
-#	""" Load data from csv or nPYc dataset object """
-#
-## 	TODO: add this functionality
-#
-#	# Create dataset
-#	class Dataset(object):
-#
-#		def __init__(self):
-#			self.intensityData = np.array(None)
-#			self.featureMetadata = pandas.DataFrame(None, columns=['Feature Name', 'Retention Time', 'm/z'])
-#
-#	self.dataset = Dataset()
-#
-#	# nPYc dataset object
-#	if (self.Attributes['nPYcDataset'] is not None):
-#
-#		self.dataset = self.Attributes['nPYcDataset']
-#		del self.Attributes['nPYcDataset']
-#
-#	# Load from csv files
-#	else:
-#
-#		self.dataset.intensityData = np.genfromtxt(self.Attributes['intensityDataFile'], delimiter=',')
-#		self.dataset.featureMetadata = pandas.read_csv(self.Attributes['featureMetadataFile'])
-#
-#	# Check attributes
-#	ds, dv = self.dataset.intensityData.shape
-#	fv = self.dataset.featureMetadata.shape[0]
-#	if dv != fv:
-#		raise ValueError('intensityData and featureMetadata have different dimensions')
+def _loadDatasetObject(self):
+	""" Load data from nPYc dataset object: note, 'nPYcDataset must be input argument on initiation of runISTOCSY """
+
+	# nPYc dataset object
+	self.dataset = self.Attributes['nPYcDataset']
+	del self.Attributes['nPYcDataset']
+
+	# Check attributes
+	ds, dv = self.dataset.intensityData.shape
+	fv = self.dataset.featureMetadata.shape[0]
+	if dv != fv:
+		raise ValueError('intensityData and featureMetadata have different dimensions')
 
 
 def _findNearest(featureMetadata, Xvalue, Yvalue):
